requirements.txt
torch>=2.0.0
transformers>=4.30.0
captum==0.7.0
pandas>=1.5.0
numpy>=1.24.0
nltk>=3.8.0
tqdm>=4.65.0
pyyaml>=6.0